refactor: extract per-branch turn queue for reuse by regenerate
